# techBuddy/src/main/java/com/amalitech/techBuddy/dto/UserDto.java
package com.amalitech.techBuddy.dto;

import com.amalitech.techBuddy.entity.Role;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class UserDto {
    private String username;
    private String password;
    private List<Role> role;
}